Use float as comparand type

+++ src/filtras/comparator.h
@@ -106,7 +106,7 @@ public:
                 {"required", true}
             }},
             {"comparand", {
-                {"type", "integer"},
+                {"type", "float"},
                 {"required", true}
             }}
         });
